+++ assets/admin.js
@@ -2,21 +2,17 @@
 	'use strict';
 
 	$(function () {
-		var $type = $('#lsar-business-type');
+		/*
+		 * Visibility comes from PHP (WooCommerce class + wc_get_products).
+		 * Product schema is not limited to business type Store.
+		 */
 		var wcActive = !!(window.lsarAdmin && window.lsarAdmin.woocommerceActive);
+		var $rows = $('.lsar-woocommerce-field').closest('tr');
 
-		function toggleWooCommerceFields() {
-			var isStore = $type.val() === 'Store';
-			var $rows = $('.lsar-woocommerce-field').closest('tr');
-
-			if (isStore && wcActive) {
-				$rows.show();
-			} else {
-				$rows.hide();
-			}
+		if (wcActive) {
+			$rows.show();
+		} else {
+			$rows.hide();
 		}
-
-		$type.on('change', toggleWooCommerceFields);
-		toggleWooCommerceFields();
 	});
 })(jQuery);